pyproject.toml
[project]
name = "easyrag"
version = "1.0.0"
description = "A simple RAG (Retrieval-Augmented Generation) API using Qdrant and LangChain"
requires-python = ">=3.11"
dependencies = [
    "langchain",
    "langchain-core",
    "langchain-community",
    "langchain-huggingface",
    "langchain-qdrant",
    "qdrant-client[grpc]",
    "pymupdf",
    "transformers",
    "torch",
    "fastapi",
    "uvicorn",
    "python-multipart",
    "pydantic-settings",
    "sentence-transformers>=5.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/easyrag"]

[tool.uv]
dev-dependencies = []
